pyproject.toml
[project]
name = "repl-nix-discord-wc3-random-voiceline-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chalkpy>=2.51.1",
    "discord-py==2.3.2",
    "flask>=3.0.3",
    "pynacl==1.5.0",
    "python-dotenv>=1.0.1",
]
